requirements/base.txt
asgiref==3.8.1
Django==5.0.4
python-dotenv==1.0.1
sqlparse==0.4.4
